use different urls instead of modals and state
