--- frontend/src/lib/components/OrderDialog.svelte
@@ -1,219 +1,0 @@
-<script lang="ts">
-	import {
-		type Equipment,
-		equipmentLabels,
-		EquipmentType,
-		randomRegistrationCode
-	} from '$lib/services/equipment.service';
-	import { createOrder, deleteOrder, fulfillOrder, getOrdersForMember, type Order } from '$lib/services/order.service';
-	import { getNextGloveId } from '$lib/services/gloveId.service';
-	import { onMount } from 'svelte';
-	import { showError } from '$lib/services/notification.svelte';
-	import { formatToDate } from '$lib/services/timeHelper';
-	import { equipmentForGroup, type Member, saveEquipmentForMember } from '$lib/services/member.service';
-
-	interface Props {
-		member: Member;
-		onClose: () => void;
-		onEquipmentChanged: (equipment: Equipment) => void;
-		onOrderCreated: (order: Order) => void;
-		onOrderRemoved: (orderId: number) => void;
-	}
-	let { member, onClose, onEquipmentChanged, onOrderCreated, onOrderRemoved }: Props = $props();
-
-	let orderSizes: Record<EquipmentType, string> = $state({
-		[EquipmentType.Helmet]: '0',
-	}) as Record<EquipmentType, string>;
-	let registrationCodes: Record<EquipmentType, string> = $state({
-		[EquipmentType.Helmet]: randomRegistrationCode()
-	}) as Record<EquipmentType, string>;
-	let orders: Order[] = $state([]);
-	let loadingOrders = $state(false);
-	let nextGloveId: string | null = $state(null);
-	let loadingGloveId = $state(false);
-
-	const loadOrders = async () => {
-		loadingOrders = true;
-		orders = await getOrdersForMember(member.id);
-		loadingOrders = false;
-	};
-
-	const loadNextGloveId = async () => {
-		try {
-			loadingGloveId = true;
-			nextGloveId = (await getNextGloveId()).nextId;
-		} catch (error) {
-			console.error('Failed to load next glove ID:', error);
-			nextGloveId = null;
-		} finally {
-			loadingGloveId = false;
-		}
-	};
-
-	onMount(() => {
-		loadOrders();
-		loadNextGloveId();
-	})
-
-	const getOpenOrder = (type: EquipmentType) =>
-		orders.find(order => order.type === type && !order.fulfilledAt);
-
-	const handleOrderEquipment = async (e: SubmitEvent,equipmentType: EquipmentType) => {
-		e.preventDefault();
-		const size = orderSizes[equipmentType]?.trim();
-		try {
-			const memberIdInternal = parseInt(member.id, 10);
-			const newOrder = await createOrder({
-				id: 0,
-				type: equipmentType,
-				createdAt: undefined,
-				size,
-				memberId: memberIdInternal,
-				fulfilledAt: undefined
-			});
-			if (equipmentType !== EquipmentType.Helmet) {
-				orderSizes[equipmentType] = '';
-			}
-			orders = [...orders, newOrder];
-			onOrderCreated(newOrder);
-		} catch (e) {
-			console.error(e);
-			showError("Fehler beim Erstellen der Bestellung.");
-		}
-	};
-
-	const handleDeleteOrder = async (orderId: number) => {
-		await deleteOrder(orderId);
-		orders = orders.filter(order => order.id !== orderId);
-		onOrderRemoved(orderId);
-	};
-
-	const handleFulfillOrder = async (e: SubmitEvent, order: Order, equipmentType: EquipmentType) => {
-		e.preventDefault()
-		const regCode = registrationCodes[equipmentType]?.trim();
-		if (!regCode) {
-			alert('Bitte eine Ausrüstungsnummer angeben.');
-			return;
-		}
-
-		try {
-			const equipment = await fulfillOrder(order, regCode);
-			registrationCodes[equipmentType] = '';
-
-			orders = orders.filter(o => order.id !== o.id);
-			onOrderRemoved(order.id);
-			onEquipmentChanged(equipment);
-		} catch (error) {
-			console.error(error);
-			showError('Fehler beim Erfüllen der Bestellung.');
-		}
-	};
-</script>
-
-	<div class="fixed inset-0 bg-black bg-opacity-40 flex items-center justify-center z-60">
-		<div class="bg-white rounded-lg p-6 w-full max-w-md max-h-[95vh] overflow-y-auto">
-			<div class="flex items-center justify-between mb-4">
-				<h3 class="text-lg font-bold">Bestellungen {member.name}</h3>
-				<button
-					type="button"
-					onclick={onClose}
-					class="text-gray-400 hover:text-gray-600 text-2xl leading-none"
-				>
-					×
-				</button>
-			</div>
-			{#if loadingOrders}
-				<div class="text-center py-4">Lade Bestellungen...</div>
-			{:else}
-				<div class="space-y-4">
-					{#each equipmentForGroup[member.group] as equipmentType}
-						{@const openOrder = getOpenOrder(equipmentType)}
-						<div class="border rounded-lg p-4 flex flex-col gap-1">
-							<div class="flex items-center gap-2">
-								<span class="w-28">{equipmentLabels[equipmentType]}</span>
-								{#if openOrder}
-									{#if equipmentType !== EquipmentType.Helmet}
-										<span class="text-xs text-gray-600 bg-gray-100 rounded px-2 py-0.5" title="Bestellte Größe">
-											Größe: {openOrder.size}
-										</span>
-									{/if}
-									<span class="text-xs text-gray-500 bg-gray-50 rounded px-2 py-0.5" title="Bestelldatum">
-										{formatToDate(openOrder.createdAt)}
-									</span>
-									<div class="flex-1"></div>
-									<button
-										type="button"
-										class="px-3 py-1 bg-red-600 text-white rounded hover:bg-red-700 transition-colors ml-auto"
-										onclick={() => handleDeleteOrder(openOrder.id)}
-									>
-										Löschen
-									</button>
-								{:else}
-									<form onsubmit={e => handleOrderEquipment(e, equipmentType)}>
-									{#if equipmentType !== EquipmentType.Helmet}
-										<input
-											type="text"
-											placeholder="Größe"
-											class="flex-1 px-2 py-1 border border-gray-300 rounded focus:border-blue-500"
-											required
-											bind:value={orderSizes[equipmentType]}
-										/>
-									{:else}
-										<div class="flex-1"></div>
-									{/if}
-									<button
-										type="submit"
-										class="px-3 py-1 bg-blue-600 text-white rounded hover:bg-blue-700 transition-colors"
-									>
-										Bestellen
-									</button>
-									</form>
-								{/if}
-							</div>
-
-							{#if openOrder}
-								<form onsubmit={e => handleFulfillOrder(e, openOrder, equipmentType)} class="flex items-center gap-2 mt-1">
-									{#if equipmentType !== EquipmentType.Helmet}
-										<input
-											type="text"
-											placeholder="Ausrüstungsnummer"
-											required
-											class="flex-1 px-2 py-1 border border-gray-300 rounded focus:border-blue-500"
-											bind:value={registrationCodes[equipmentType]}
-										/>
-									{:else}
-										<div class="flex-1"></div>
-									{/if}
-									<button
-										type="submit"
-										class="px-3 py-1 bg-green-600 text-white rounded hover:bg-green-700 transition-colors"
-									>
-										Erfüllen
-									</button>
-								</form>
-								{#if equipmentType === EquipmentType.Gloves && nextGloveId}
-									<div class="text-xs text-green-600 mt-1">
-										✓ Nächste verfügbare Handschuh-ID: {nextGloveId}
-									</div>
-								{/if}
-								{#if equipmentType === EquipmentType.Gloves && loadingGloveId}
-									<div class="text-xs text-blue-600 mt-1">
-										⏳ Lade nächste Handschuh-ID...
-									</div>
-								{/if}
-							{/if}
-						</div>
-					{/each}
-				</div>
-			{/if}
-			<div class="flex justify-end mt-6">
-				<button
-					type="button"
-					onclick={onClose}
-					class="px-4 py-2 text-gray-600 bg-gray-200 rounded hover:bg-gray-300 transition-colors"
-				>
-					Schließen
-				</button>
-			</div>
-		</div>
-	</div>
