+++ frontend/svelte.config.js
@@ -11,7 +11,7 @@ const config = {
 		// adapter-auto only supports some environments, see https://svelte.dev/docs/kit/adapter-auto for a list.
 		// If your environment is not supported, or you settled on a specific environment, switch out the adapter.
 		// See https://svelte.dev/docs/kit/adapters for more information about adapters.
-		adapter: adapter({}) }
+		adapter: adapter({ fallback: 'index.html' }) }
 };
 
 export default config;
